refactor(auth): extract admin-helpers to src/lib/auth/admin-helpers.ts (CP-4)
