+++ src/lib/auth/admin-helpers.ts
@@ -1,0 +1,30 @@
+import { env } from "@/env";
+import { sendWelcomeEmail } from "@/lib/email";
+import { logger } from "@/lib/logger";
+
+export function getAdminEmails(): {
+  superAdmins: string[];
+  admins: string[];
+} {
+  const superAdmins = env.SUPER_ADMIN_EMAILS.split(",").filter(Boolean);
+  const admins = env.ADMIN_EMAILS.split(",").filter(Boolean);
+  return { superAdmins, admins };
+}
+
+export async function handleWelcomeEmail(user: {
+  email: string;
+  name: string;
+}): Promise<void> {
+  try {
+    await sendWelcomeEmail({
+      to: user.email,
+      userName: user.name,
+    });
+  } catch (error) {
+    logger.error({
+      type: "email:welcome:failed",
+      email: user.email,
+      error: error instanceof Error ? error.message : String(error),
+    });
+  }
+}
